chore: vote 패키지 분리
- board 패키지 생성
- vote 내에 expireVote 관련 기능을 board 패키지로 분리
