+++ src/main/java/ssafy/lambda/board/entity/VoteComment.java
@@ -1,4 +1,4 @@
-package ssafy.lambda.vote.entity;
+package ssafy.lambda.board.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
