+++ src/main/java/ssafy/lambda/vote/controller/VoteController.java
@@ -5,7 +5,6 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import ssafy.lambda.vote.dto.RequestVoteDto;
-import ssafy.lambda.vote.dto.ResponseCommentDto;
 import ssafy.lambda.vote.dto.ResponseProfileWithPercentDto;
 import ssafy.lambda.vote.dto.ResponseVoteDto;
 import ssafy.lambda.vote.dto.ResponseVoteStatusDto;
@@ -24,7 +22,6 @@ import ssafy.lambda.vote.service.VoteService;
 @Slf4j
 public class VoteController {
 
-    // @Autowired
     private final VoteService voteService;
 
     @Operation(summary = "투표 생성", description = "유저가 새로운 투표를 만듭니다")
@@ -106,41 +103,4 @@ public class VoteController {
                              .body(responseVoteStatusDto);
     }
 
-
-    @Operation(summary = "댓글 조회", description = "만료된 투표에 대한 댓글 목록을 반환합니다")
-    @GetMapping("/comment")
-    public ResponseEntity<List<ResponseCommentDto>> getComments(
-        @RequestParam Long voteId
-    ) {
-        log.info("getVoteResult - vote {} ", voteId);
-        List<ResponseCommentDto> voteResult = voteService.getComments(voteId);
-        return ResponseEntity.ok()
-                             .body(voteResult);
-    }
-
-    @Operation(summary = "댓글 작성", description = "만료된 투표에 대한 댓글을 작성합니다")
-    @PostMapping("/comment")
-    public ResponseEntity writeComment(
-        @RequestParam Long voteId,
-        @RequestParam Long memberId,
-        @RequestBody String content
-    ) {
-        log.info("writeComment - member {}, vote {}  : {}", memberId, voteId, content);
-        voteService.writeComment(voteId, memberId, content);
-        return ResponseEntity.ok()
-                             .build();
-    }
-
-
-    @Operation(summary = "댓글 삭제", description = "작성한 댓글을 삭제합니다")
-    @DeleteMapping("/comment/{commentId}")
-    public ResponseEntity deleteComment(
-        @PathVariable Long commentId,
-        @RequestParam Long memberId
-    ) {
-        log.info("deleteComment - member {}, comment {}  : {}", memberId, commentId);
-        voteService.deleteComment(commentId);
-        return ResponseEntity.ok()
-                             .build();
-    }
 }
