+++ src/main/java/ssafy/lambda/vote/service/VoteService.java
@@ -2,7 +2,6 @@ package ssafy.lambda.vote.service;
 
 import java.util.List;
 import ssafy.lambda.vote.dto.RequestVoteDto;
-import ssafy.lambda.vote.dto.ResponseCommentDto;
 import ssafy.lambda.vote.dto.ResponseProfileWithPercentDto;
 import ssafy.lambda.vote.dto.ResponseVoteDto;
 import ssafy.lambda.vote.dto.ResponseVoteStatusDto;
@@ -20,13 +19,6 @@ public interface VoteService {
     List<ResponseVoteDto> getUserVote(Long memberId, Long teamId);
 
     ResponseVoteStatusDto sortByVoteStatus(Long memberId, List<Long> teamIds);
-
-
-    List<ResponseCommentDto> getComments(Long expriedVoteId);
-
-    void writeComment(Long voteId, Long memberId, String content);
-
-    void deleteComment(Long commentId);
 
 
 }
